Fixed unloadImage, added "ANFT" support
ANFT stands for Animated Nitrogenfingers Text, and is identical to NFT but has a special character ("\29") to differentiate multiple layers.
ANFT loaded images are just regular NFT images put into a table. loadImage and unloadImage will treat them just 
loadImage used to add an extra line on the right, but now it's fixed.
unloadImage used to ignore " " when determining color (which is terrible, how long has that been the case!?), but that's fixed.
NFTE can now be used both by os.loadAPI() AND with require/dofile. I just want to get rid of that snippet nonsense.