Added NFP loading support in loadImage
I also fixed up drawImageCenter and drawImageCenterTransparent, which would draw the image one pixel to the left and above where it should've.
I also added stretchImageKeepAspect(), which stretches an image, but keeps the aspect ratio!  Basically, think of it fitting the image to a box with the width/height you input.